+++ lib/src/widgets/const_widgets.dart
@@ -4,6 +4,7 @@
 export 'scaled_padding.dart';
 export 'scaled_margin.dart';
 export 'scaled_container.dart';
+export 'scaled_text.dart';
 export 'spacing_widgets.dart';
 export 'const_widget_factory.dart';
 export 'skit.dart';